+++ commons_android/src/test/java/com/github/barriosnahuel/vossosunboton/commons/android/analytics/AnalyticsCoverageMatrixTest.kt
@@ -96,6 +96,7 @@ internal class AnalyticsCoverageMatrixTest {
                 "ImportHubOpened",
                 "ImportHubImportSelected",
                 "ImportHubRecordSelected",
+                "ImportHubBringSelected",
                 "RecordingCompleted",
                 "RecordPermissionResult",
                 "RecordingDraftBannerShown",